mqtt.py
import paho.mqtt.client as mqtt
from settings import Settings

settings = Settings()

def on_connect(client, userdata, flags, rc):
    print("Connected with result code "+str(rc))
    client.subscribe("sda/2022")

def on_message(client, userdata, msg):
    print(msg.topic+" "+str(msg.payload))

client = mqtt.Client()
client.on_connect = on_connect
client.on_message = on_message

client.connect(settings.MQTT_HOST, 1883, 60)

client.loop_forever()