requirements.txt
pyserial==3.5
pydantic==1.10.2
python-dotenv==0.21.0
black==22.8.0
uvicorn==0.18.3
fastapi==0.85.0
psycopg2-binary==2.9.2